Separate namespace and ontology information
A Namespace is a container for Entity objects but is not an Entity (and
not an rdf:Resource) itself. A Namespace may have an Ontology, which is
an Entity, but either way it may also contain Class, Property and
Entity objects.

All the find_entity() etc. functions have moved to Namespace.