Intern class properties - they are now indexed per class.
This presents an issue when a resource has more than one rdf:type. The
sensible solution I think is to create a psuedoclass that is a union of
all the classes that make up its type, and this psuedoclass will have its
own index of properties. The thinking here is that the combination space
of classes is not very large, because most don't make sense together and
there will probably be many resources for each specific combination that
is used.